Add pizza SVG icon and update loading indicator
Added a new pizza SVG icon and created a `PizzaLoading` widget to use it as a custom loading indicator. Updated the `AppMessages` class to optionally use the `PizzaLoading` widget in the loading dialog. Modified the `ExploreFoodsScreen` and `SplashScreen` to use the new `PizzaLoading` widget for loading states. Also, updated `pubspec.yaml` and `pubspec.lock` to include the `flutter_svg` dependency.